RunTwenty fix

+++ HemaDungeon/Models/StateModel.cs
@@ -8,7 +8,7 @@ public sealed class StateModel
 
     public int Abdominal { get; set; }
 
-    public int RunTwenty { get; set; }
+    public float RunTwenty { get; set; }
 
     public int Rang { get; set; }
